fix output of .list_experiments() to only output lists of IDs

+++ shepherd_client/examples/e5_experiment_download_and_delete.py
@@ -17,8 +17,10 @@ for xp_id in client.list_experiments(only_finished=True):
     client.download_experiment(xp_id, path_here)
 
 # and later when you are sure your data is fine:
-for xp_id in client.list_experiments(only_finished=True):
-    client.delete_experiment(xp_id)
+for xp_id in client.list_experiments():
+    state = client.get_experiment_state(xp_id)
+    if state in {"failed", "finished"}:
+        client.delete_experiment(xp_id)
 # end example
 
 # ########### EXTRA ##################
